+++ attachments/src/main/java/com/backstage/app/attachment/configuration/properties/AttachmentProperties.java
@@ -24,6 +24,7 @@ import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.http.MediaType;
 
+import java.time.Duration;
 import java.util.Set;
 
 @Getter
@@ -149,5 +150,10 @@ public class AttachmentProperties
 	/**
 	 * Активирует механизм удаления не привязанных вложений.
 	 */
-	private boolean deleteUnbounded;
+	private boolean deleteUnbound;
+
+	/**
+	 * Интервал, после которого не привязанное вложение будет удалено.
+	 */
+	private Duration deleteUnboundDuration = Duration.ofDays(1);
 }