+++ attachments/src/main/java/com/backstage/app/attachment/repository/AttachmentRepository.java
@@ -31,5 +31,11 @@ import java.util.List;
 public interface AttachmentRepository extends CustomJpaRepository<Attachment, String>
 {
 	@Query("select distinct a.id from Attachment a left outer join a.bindings ab where ab.attachment is null and a.created < :date")
-	List<String> findExpiredUnbounded(@Param("date") LocalDateTime expirationDate, Pageable pageable);
+	List<String> findExpiredUnbound(@Param("date") LocalDateTime expirationDate, Pageable pageable);
+
+	@Query("select distinct a.id from Attachment a left outer join a.bindings ab where ab.attachment is null")
+	Long countUnbound();
+
+	@Query("select sum(a.size) from Attachment a")
+	Long countTotalSize();
 }